+++ airbyte-config/init/src/main/resources/config/STANDARD_SOURCE_DEFINITION/9845d17a-45f1-4070-8a60-50914b1c8e2b.json
@@ -1,7 +1,7 @@
 {
   "sourceDefinitionId": "9845d17a-45f1-4070-8a60-50914b1c8e2b",
-  "name": "REST API",
-  "dockerRepository": "airbyte/source-rest-api",
+  "name": "HTTP Request",
+  "dockerRepository": "airbyte/source-http-request",
   "dockerImageTag": "0.1.0",
-  "documentationUrl": "https://hub.docker.com/r/airbyte/source-rest-api"
+  "documentationUrl": "https://hub.docker.com/r/airbyte/source-http-request"
 }
