+++ airbyte-integrations/connectors/source-http-request/Dockerfile
@@ -2,9 +2,9 @@ FROM airbyte/integration-base-python:dev
 
 RUN apt-get update && apt-get install -y jq curl bash && rm -rf /var/lib/apt/lists/*
 
-ENV CODE_PATH="source_rest_api"
-ENV AIRBYTE_IMPL_MODULE="source_rest_api"
-ENV AIRBYTE_IMPL_PATH="SourceRestApi"
+ENV CODE_PATH="source_http_request"
+ENV AIRBYTE_IMPL_MODULE="source_http_request"
+ENV AIRBYTE_IMPL_PATH="SourceHttpRequest"
 
 WORKDIR /airbyte/integration_code
 COPY $CODE_PATH ./$CODE_PATH
@@ -12,4 +12,4 @@ COPY setup.py ./
 RUN pip install ".[main]"
 
 LABEL io.airbyte.version=0.1.0
-LABEL io.airbyte.name=airbyte/source-rest-api
+LABEL io.airbyte.name=airbyte/source-http-request
