+++ airbyte-integrations/connectors/source-http-request/Dockerfile.test
@@ -4,16 +4,16 @@ RUN apt-get update && rm -rf /var/lib/apt/lists/*
 
 ENV CODE_PATH="integration_tests"
 ENV AIRBYTE_TEST_MODULE="integration_tests"
-ENV AIRBYTE_TEST_PATH="SourceRestApiStandardTest"
+ENV AIRBYTE_TEST_PATH="SourceHttpRequestStandardTest"
 ENV AIRBYTE_TEST_CASE=true
 
 LABEL io.airbyte.version=0.1.0
-LABEL io.airbyte.name=airbyte/source-rest-api-standard-test
+LABEL io.airbyte.name=airbyte/source-http-request-standard-test
 
 WORKDIR /airbyte/integration_code
-COPY source_rest_api source_rest_api
+COPY source_http_request source_http_request
 COPY $CODE_PATH $CODE_PATH
-COPY source_rest_api/*.json $CODE_PATH
+COPY source_http_request/*.json $CODE_PATH
 COPY setup.py ./
 
 RUN pip install ".[tests]"
