+++ airbyte-integrations/connectors/source-http-request/integration_tests/catalog.json
