+++ airbyte-integrations/connectors/source-http-request/requirements.txt
