+++ airbyte-integrations/connectors/source-http-request/source_http_request/source.py
@@ -31,7 +31,7 @@ from airbyte_protocol import AirbyteCatalog, AirbyteConnectionStatus, AirbyteMes
 from base_python import AirbyteLogger, Source
 
 
-class SourceRestApi(Source):
+class SourceHttpRequest(Source):
     STREAM_NAME = "data"
 
     def __init__(self):
@@ -54,7 +54,7 @@ class SourceRestApi(Source):
         }
 
         # json body will be returned as the "data" stream". we can't know its schema ahead of time, so we assume it's object (i.e. valid json).
-        return AirbyteCatalog(streams=[AirbyteStream(name=SourceRestApi.STREAM_NAME, json_schema=json_schema)])
+        return AirbyteCatalog(streams=[AirbyteStream(name=SourceHttpRequest.STREAM_NAME, json_schema=json_schema)])
 
     def read(self, logger: AirbyteLogger, config_container, catalog_path, state=None) -> Generator[AirbyteMessage, None, None]:
         r = self._make_request(config_container.rendered_config)
@@ -64,8 +64,11 @@ class SourceRestApi(Source):
         # need to eagerly fetch the json.
         message = AirbyteMessage(
             type=Type.RECORD,
-            record=AirbyteRecordMessage(stream=SourceRestApi.STREAM_NAME, data=r.json(), emitted_at=int(datetime.now().timestamp()) * 1000),
+            record=AirbyteRecordMessage(
+                stream=SourceHttpRequest.STREAM_NAME, data=r.json(), emitted_at=int(datetime.now().timestamp()) * 1000
+            ),
         )
+
         return (m for m in [message])
 
     def _make_request(self, config):
