+++ airbyte-integrations/connectors/source-http-request/source_http_request/spec.json
