+++ airbyte-integrations/connectors/source-http-request/unit_tests/unit_test.py
@@ -22,6 +22,26 @@ OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
 SOFTWARE.
 """
 
-from .source import SourceRestApi
+import unittest
 
-__all__ = ["SourceRestApi"]
+from source_http_request import SourceHttpRequest
+
+
+class TestSourceHttpRequest(unittest.TestCase):
+    def test_parse_config(self):
+        config = {
+            "http_method": "get",
+            "url": "http://api.bart.gov/api",
+            "headers": '{"Content-Type": "application/json"}',
+            "body": '{"something": "good"}',
+        }
+
+        source = SourceHttpRequest()
+        actual = source._parse_config(config)
+        expected = {
+            "http_method": "get",
+            "url": "http://api.bart.gov/api",
+            "headers": {"Content-Type": "application/json"},
+            "body": {"something": "good"},
+        }
+        self.assertEqual(expected, actual)
