+++ docs/SUMMARY.md
@@ -18,13 +18,13 @@
     * [Google Analytics](integrations/sources/googleanalytics.md)
     * [Google Sheets](integrations/sources/google-sheets.md)
     * [Hubspot](integrations/sources/hubspot.md)
+    * [HTTP Request](integrations/sources/http-request.md)
     * [Mailchimp](integrations/sources/mailchimp.md)
     * [Marketo](integrations/sources/marketo.md)
     * [MySQL](integrations/sources/mysql.md)
     * [Microsoft SQL Server \(MSSQL\)](integrations/sources/mssql.md)
     * [Postgres](integrations/sources/postgres.md)
     * [Recurly](integrations/sources/recurly.md)
-    * [Rest API](integrations/sources/rest-api.md)
     * [Salesforce](integrations/sources/salesforce.md)
     * [Sendgrid](integrations/sources/sendgrid.md)
     * [Shopify](integrations/sources/shopify.md)
