+++ docs/integrations/sources/http-request.md
@@ -1,4 +1,4 @@
-# Rest API
+# HTTP Request
 
 ## Overview
